requirement.txt
django
xlsxwriter
django-crispy-forms
django-tables2
docxcompose
docxtpl
django_crispy_forms
django_filter
django_htmx
psycopg2
requests
django_static_jquery
jinja2
bcrypt
django-cors-headers